Adjusted folder hierarchy to match `main` schema
