+++ src/components/Chart/Chart.js
@@ -21,12 +21,12 @@ import cn from "classnames";
 import { Line } from "react-chartjs-2";
 
 import PeriodPicker from "./PeriodPicker";
-import formatDatePeriod from "../../../utils/formatDatePeriod";
+import formatDatePeriod from "../../utils/formatDatePeriod";
 import calcMetricPercentage from "./utils/calcMetricPercentage";
 import adjustChartDataLength from "./utils/adjustChartDataLength";
 import { chartDataPropTypes } from "./propTypes";
-import formatNumber from "../../../utils/formatNumber";
-import months from "../../../constants/months";
+import formatNumber from "../../utils/formatNumber";
+import months from "../../constants/months";
 
 import "./Chart.scss";
