+++ src/components/Chart/PeriodPicker.js
