+++ src/components/Chart/PeriodPicker.scss
@@ -37,7 +37,7 @@
       display: block;
       width: 0.6875rem;
       height: 0.3125rem;
-      background: url("../../../assets/icons/select-arrow.svg") center center no-repeat;
+      background: url("../../assets/icons/select-arrow.svg") center center no-repeat;
       background-size: cover;
     }
     &--opened {
