+++ src/components/Chart/__tests__/PeriodPicker.test.js
