+++ src/components/Chart/propTypes.js
