+++ src/components/Chart/utils/__tests__/adjustChartDataLength.test.js
