+++ src/components/Chart/utils/__tests__/calcMetricPercentage.js
@@ -14,31 +14,23 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-import React from "react";
+import calcMetricPercentage, { NO_DATA_ERROR } from "../calcMetricPercentage";
+import logger from "../../../../utils/logger";
 
-import Card from "../Card";
-import { keyInsightsPropTypes } from "./propTypes";
+describe("calcPercentage.js", () => {
+  const logErrorSpy = jest.spyOn(logger, "error");
 
-import "./KeyInsights.scss";
+  it("should calculate percentage from chart dataset", () => {
+    expect(calcMetricPercentage([100, null, 95, 120, 75, 80, 155])).toBe("+55%");
+    expect(calcMetricPercentage([100, 95, 90, 80, 75, 70])).toBe("-30%");
+  });
 
-const KeyInsights = ({ keyInsightsData }) => (
-  <div className="KeyInsights">
-    <h2 className="KeyInsights__title">Key Insights</h2>
-    <div className="KeyInsights__cards">
-      {keyInsightsData.map((card) => (
-        <div key={card.title} className="KeyInsights__card">
-          <Card {...card} />
-          <p className="KeyInsights__card-description">
-            {card.caption} {card.reportingCountiesModal}
-          </p>
-        </div>
-      ))}
-    </div>
-  </div>
-);
+  it("should throw error to console when data array is empty", () => {
+    expect(calcMetricPercentage([])).toBe("N/A");
+    expect(logErrorSpy).toBeCalledWith(NO_DATA_ERROR);
+  });
 
-KeyInsights.propTypes = {
-  keyInsightsData: keyInsightsPropTypes.isRequired,
-};
-
-export default KeyInsights;
+  it("should work when array consists of the only data point", () => {
+    expect(calcMetricPercentage([1])).toBe("0%");
+  });
+});
