+++ src/components/Chart/utils/adjustChartDataLength.js
