+++ src/components/Chart/utils/calcMetricPercentage.js
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
-import logger from "../../../../utils/logger";
+import logger from "../../../utils/logger";
 
 export const NO_DATA_ERROR = "Cannot generate percentage for empty metric.";
