+++ src/components/Corrections/Corrections.js
@@ -18,14 +18,14 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import FlowDiagram from "./FlowDiagram";
-import KeyInsights from "../shared/KeyInsights";
+import KeyInsights from "../KeyInsights";
 import ErrorBoundary from "../shared/ErrorBoundary";
-import Chart from "../shared/Chart";
-import Sources from "../shared/Sources";
-import { chartDataPropTypes } from "../shared/Chart/propTypes";
+import Chart from "../Chart";
+import Sources from "../Sources";
+import { chartDataPropTypes } from "../Chart/propTypes";
 import { flowDiagramDataPropTypes } from "./FlowDiagram/propTypes";
-import { keyInsightsPropTypes } from "../shared/KeyInsights/propTypes";
-import { sourcePropTypes } from "../shared/Sources/propTypes";
+import { keyInsightsPropTypes } from "../KeyInsights/propTypes";
+import { sourcePropTypes } from "../Sources/propTypes";
 
 const Corrections = ({
   populationsChartData,
